+++ squadcastv1/src/squadcast_sdk/workflows_sdk.py
@@ -2808,7 +2808,7 @@ class WorkflowsSDK(BaseSDK):
         server_url: Optional[str] = None,
         timeout_ms: Optional[int] = None,
         http_headers: Optional[Mapping[str, str]] = None,
-    ) -> models.V3WorkflowsActionResponse:
+    ) -> models.WorkflowsUpdateWorkflowActionResponse:
         r"""Update Workflow Action
 
         Update an action by action ID
@@ -2901,7 +2901,9 @@ class WorkflowsSDK(BaseSDK):
 
         response_data: Any = None
         if utils.match_response(http_res, "200", "application/json"):
-            return unmarshal_json_response(models.V3WorkflowsActionResponse, http_res)
+            return unmarshal_json_response(
+                models.WorkflowsUpdateWorkflowActionResponse, http_res
+            )
         if utils.match_response(http_res, "400", "application/json"):
             response_data = unmarshal_json_response(
                 errors.BadRequestErrorData, http_res
@@ -2973,7 +2975,7 @@ class WorkflowsSDK(BaseSDK):
         server_url: Optional[str] = None,
         timeout_ms: Optional[int] = None,
         http_headers: Optional[Mapping[str, str]] = None,
-    ) -> models.V3WorkflowsActionResponse:
+    ) -> models.WorkflowsUpdateWorkflowActionResponse:
         r"""Update Workflow Action
 
         Update an action by action ID
@@ -3066,7 +3068,9 @@ class WorkflowsSDK(BaseSDK):
 
         response_data: Any = None
         if utils.match_response(http_res, "200", "application/json"):
-            return unmarshal_json_response(models.V3WorkflowsActionResponse, http_res)
+            return unmarshal_json_response(
+                models.WorkflowsUpdateWorkflowActionResponse, http_res
+            )
         if utils.match_response(http_res, "400", "application/json"):
             response_data = unmarshal_json_response(
                 errors.BadRequestErrorData, http_res