+++ squadcastv1/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [python v0.3.5] squadcastv1
 ### Releases
 - [PyPI v0.3.5] https://pypi.org/project/squadcast_sdk/0.3.5 - squadcastv1
+
+## 2026-02-04 00:06:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.701.0 (2.803.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.5.0] squadcastv1
+### Releases
+- [PyPI v0.5.0] https://pypi.org/project/squadcast_sdk/0.5.0 - squadcastv1